Improved (more organized) map linking handling, some level generation code refinements, and updated tests

+++ tests/code/locationTest.cpp
@@ -4,7 +4,7 @@
 
 using namespace std;
 
-int IMapTest::generate()
+long int IMapTest::generate()
 {
     IUnitTest * unitTestHarness = locator::getUnitTests();
